Refactor submitJob out in a separate service
