+++ src/main/scala/za/co/absa/hyperdrive/trigger/configuration/application/GeneralConfig.scala
@@ -35,5 +35,7 @@ class GeneralConfig (
   @DefaultValue(Array("Unknown"))
   val version: String,
   @(NotBlank @field)
-  val appUniqueId: String
+  val appUniqueId: String,
+  @Name("spark.cluster.api")
+  val sparkClusterApi: String
 )
