+++ src/main/scala/za/co/absa/hyperdrive/trigger/scheduler/executors/Executors.scala
@@ -22,11 +22,12 @@ import za.co.absa.hyperdrive.trigger.models.{DagInstance, JobInstance, ShellInst
 import za.co.absa.hyperdrive.trigger.models.enums.JobStatuses.InvalidExecutor
 import za.co.absa.hyperdrive.trigger.models.enums.{DagInstanceStatuses, JobStatuses}
 import za.co.absa.hyperdrive.trigger.persistance.{DagInstanceRepository, JobInstanceRepository}
-import za.co.absa.hyperdrive.trigger.scheduler.executors.spark.SparkExecutor
+import za.co.absa.hyperdrive.trigger.scheduler.executors.spark.{SparkClusterService, SparkExecutor, SparkYarnClusterServiceImpl}
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.BeanFactory
 import za.co.absa.hyperdrive.trigger.scheduler.executors.shell.ShellExecutor
 import org.springframework.stereotype.Component
-import za.co.absa.hyperdrive.trigger.configuration.application.{SchedulerConfig, SparkYarnSinkConfig}
+import za.co.absa.hyperdrive.trigger.configuration.application.{GeneralConfig, SchedulerConfig, SparkYarnSinkConfig}
 import za.co.absa.hyperdrive.trigger.scheduler.notifications.NotificationSender
 
 import scala.concurrent.{ExecutionContext, ExecutionContextExecutor, Future}
@@ -34,12 +35,22 @@ import scala.util.{Failure, Success}
 
 @Component
 class Executors @Inject()(dagInstanceRepository: DagInstanceRepository, jobInstanceRepository: JobInstanceRepository,
-                          notificationSender: NotificationSender, sparkYarnSinkConfig: SparkYarnSinkConfig,
-                          schedulerConfig: SchedulerConfig) {
+                          notificationSender: NotificationSender, beanFactory: BeanFactory,
+                          sparkYarnSinkConfig: SparkYarnSinkConfig, schedulerConfig: SchedulerConfig,
+                          generalConfig: GeneralConfig) {
   private val logger = LoggerFactory.getLogger(this.getClass)
   private implicit val executorConfig: ExecutorConfig = new ExecutorConfig(sparkYarnSinkConfig)
   private implicit val executionContext: ExecutionContextExecutor =
     ExecutionContext.fromExecutor(concurrent.Executors.newFixedThreadPool(schedulerConfig.executors.threadPoolSize))
+  private val sparkClusterService: SparkClusterService = {
+    generalConfig.sparkClusterApi.toLowerCase match {
+      case "yarn" => {
+        logger.info(s"Using yarn cluster")
+        beanFactory.getBean(classOf[SparkYarnClusterServiceImpl])
+      }
+      case _ => throw new IllegalArgumentException("Invalid spark cluster api - use one of: yarn")
+    }
+  }
 
   def executeDag(dagInstance: DagInstance): Future[Unit] = {
     jobInstanceRepository.getJobInstances(dagInstance.id).flatMap {
@@ -71,7 +82,7 @@ class Executors @Inject()(dagInstanceRepository: DagInstanceRepository, jobInsta
         val fut = dagInstanceRepository.update(dagInstance.copy(status = DagInstanceStatuses.Running)).flatMap { _ =>
           jobInstance match {
             case Some(ji) => ji.jobParameters match {
-              case spark: SparkInstanceParameters => SparkExecutor.execute(ji, spark, updateJob)
+              case spark: SparkInstanceParameters => SparkExecutor.execute(ji, spark, updateJob, sparkClusterService)
               case shell: ShellInstanceParameters => ShellExecutor.execute(ji, shell, updateJob)
               case _ => updateJob(ji.copy(jobStatus = InvalidExecutor))
             }
